test(billing): 💍 rename dfu-flash test files to billing-import
