+++ server/tests/integration/billing/dfu/billing-import/billing-import-cross-processor-balances.test.ts
@@ -10,7 +10,7 @@ import {
 	type FlashClient,
 	callFlash,
 	createRealStripeSub,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { expectBalanceCorrect } from "@tests/integration/utils/expectBalanceCorrect.js";
 import { TestFeature } from "@tests/setup/v2Features.js";
 import { items } from "@tests/utils/fixtures/items.js";
