+++ server/tests/integration/billing/dfu/billing-import/billing-import-dry-run.test.ts
@@ -1,0 +1,61 @@
+/**
+ * dfu.flash — dry_run: returns customer: null and persists nothing.
+ */
+
+import { expect, test } from "bun:test";
+import { type ApiCustomerV5 } from "@autumn/shared";
+import {
+	type FlashClient,
+	callFlash,
+	createRealStripeSub,
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
+import { expectCustomerProducts } from "@tests/integration/billing/utils/expectCustomerProductCorrect.js";
+import { items } from "@tests/utils/fixtures/items.js";
+import { products } from "@tests/utils/fixtures/products.js";
+import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
+import chalk from "chalk";
+
+test.concurrent(
+	`${chalk.yellowBright("dfu.flash: dry_run returns customer null and writes nothing")}`,
+	async () => {
+		const customerId = "dfu-flash-dry-run-customer";
+		const pro = products.pro({
+			id: "dfu-dry-run-pro",
+			items: [items.monthlyMessages({ includedUsage: 100 })],
+		});
+
+		const { autumnV2_3, ctx } = await initScenario({
+			customerId,
+			setup: [s.customer({ testClock: false }), s.products({ list: [pro] })],
+			actions: [],
+		});
+
+		const { customerId: stripeCustomerId, subscriptionId } =
+			await createRealStripeSub(ctx, { email: `${customerId}@example.com` });
+
+		const payload = {
+			customer_id: customerId,
+			dry_run: true,
+			processors: [{ type: "stripe", id: stripeCustomerId }],
+			billables: [
+				{
+					processor: "stripe",
+					link: { subscription_id: subscriptionId },
+					phases: [
+						{
+							starts_at: "now",
+							plans: [{ plan_id: pro.id, status: "active" }],
+						},
+					],
+				},
+			],
+		};
+
+		const flashRes = await callFlash(autumnV2_3 as FlashClient, payload);
+		expect(flashRes.result?.customer).toBeNull();
+
+		// Nothing persisted: the plan is not present on a fresh fetch.
+		const customer = await autumnV2_3.customers.get<ApiCustomerV5>(customerId);
+		await expectCustomerProducts({ customer, notPresent: [pro.id] });
+	},
+);
