+++ server/tests/integration/billing/dfu/billing-import/billing-import-hydration-canceling.test.ts
@@ -10,7 +10,7 @@ import {
 	type FlashClient,
 	callFlash,
 	getFlashedCustomerProduct,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { items } from "@tests/utils/fixtures/items.js";
 import { products } from "@tests/utils/fixtures/products.js";
 import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
