+++ server/tests/integration/billing/dfu/billing-import/billing-import-lifetime-multiline.test.ts
@@ -9,7 +9,7 @@ import { type ApiCustomerV5, ResetInterval } from "@autumn/shared";
 import {
 	type FlashClient,
 	callFlash,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { expectCustomerProducts } from "@tests/integration/billing/utils/expectCustomerProductCorrect.js";
 import { expectBalanceCorrect } from "@tests/integration/utils/expectBalanceCorrect.js";
 import { TestFeature } from "@tests/setup/v2Features.js";
