+++ server/tests/integration/billing/dfu/billing-import/billing-import-plan-shorthand.test.ts
@@ -1,6 +1,6 @@
 /**
- * dfu.flash — expired plan (contract 6, access-leak guard): status=expired is
- * reported AND the customer has no access to the plan's feature.
+ * dfu.flash — `plan` shorthand: a singular `plan` on a billable is normalized to
+ * phases:[{ starts_at:"now", plans:[plan] }] and imports the same cusProduct.
  */
 
 import { expect, test } from "bun:test";
@@ -9,8 +9,9 @@ import {
 	type FlashClient,
 	callFlash,
 	createRealStripeSub,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { expectCustomerProducts } from "@tests/integration/billing/utils/expectCustomerProductCorrect.js";
+import { expectBalanceCorrect } from "@tests/integration/utils/expectBalanceCorrect.js";
 import { TestFeature } from "@tests/setup/v2Features.js";
 import { items } from "@tests/utils/fixtures/items.js";
 import { products } from "@tests/utils/fixtures/products.js";
@@ -18,11 +19,11 @@ import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
 import chalk from "chalk";
 
 test.concurrent(
-	`${chalk.yellowBright("dfu.flash: expired plan is status=expired and grants no access")}`,
+	`${chalk.yellowBright("dfu.flash: `plan` shorthand imports same cusProduct as phases form")}`,
 	async () => {
-		const customerId = "dfu-flash-expired";
+		const customerId = "dfu-flash-plan-shorthand";
 		const pro = products.pro({
-			id: "dfu-expired-pro",
+			id: "dfu-plan-shorthand-pro",
 			items: [items.monthlyMessages({ includedUsage: 100 })],
 		});
 
@@ -42,22 +43,27 @@ test.concurrent(
 				{
 					processor: "stripe",
 					link: { subscription_id: subscriptionId },
-					// `plan` shorthand — equivalent to phases:[{ starts_at:"now", plans:[plan] }].
-					plan: { plan_id: pro.id, status: "expired" },
+					// Singular `plan` — normalized to phases:[{ starts_at:"now", plans:[plan] }].
+					plan: {
+						plan_id: pro.id,
+						status: "active",
+						balances: [{ feature_id: TestFeature.Messages, usage: 40 }],
+					},
 				},
 			],
 		};
 
 		const flashRes = await callFlash(autumnV2_2 as FlashClient, payload);
+		expect(flashRes.errorCode).not.toBe("invalid_inputs");
+		expect(flashRes.errorCode).not.toBe("invalid_request");
 
-		// ── Contract 6a: flashed status reported as expired. ──
-		const flashed = flashRes.result?.flashed?.find((f) => f.plan_id === pro.id);
-		expect(flashed?.status).toBe("expired");
-
-		// ── Contract 6b: customer has NO access to the expired plan's feature. ──
 		const customer = await autumnV2_3.customers.get<ApiCustomerV5>(customerId);
-		await expectCustomerProducts({ customer, notPresent: [pro.id] });
-		const messagesBalance = customer.balances?.[TestFeature.Messages];
-		expect(messagesBalance?.remaining ?? 0).toBe(0);
+		await expectCustomerProducts({ customer, active: [pro.id] });
+		expectBalanceCorrect({
+			customer,
+			featureId: TestFeature.Messages,
+			remaining: 60,
+			usage: 40,
+		});
 	},
 );
