+++ server/tests/integration/billing/dfu/billing-import/billing-import-revenuecat-expired.test.ts
@@ -16,7 +16,7 @@ import {
 	mockProduct,
 	mockSubscription,
 	rcSubscriptionProduct,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { expectCustomerProducts } from "@tests/integration/billing/utils/expectCustomerProductCorrect.js";
 import { TestFeature } from "@tests/setup/v2Features.js";
 import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
