+++ server/tests/integration/billing/dfu/billing-import/billing-import-revenuecat-oneoff.test.ts
@@ -15,7 +15,7 @@ import {
 	mockProduct,
 	mockPurchase,
 	rcOneOffProduct,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
 import chalk from "chalk";
 import { RCMappingService } from "@/external/revenueCat/misc/RCMappingService.js";
