+++ server/tests/integration/billing/dfu/billing-import/billing-import-revenuecat-seed.test.ts
@@ -12,7 +12,7 @@ import {
 	mockProduct,
 	mockSubscription,
 	rcSubscriptionProduct,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
 import chalk from "chalk";
 import { eq } from "drizzle-orm";
