+++ server/tests/integration/billing/dfu/billing-import/billing-import-stripe-recurring.test.ts
@@ -1,6 +1,7 @@
 /**
- * dfu.flash — `plan` shorthand: a singular `plan` on a billable is normalized to
- * phases:[{ starts_at:"now", plans:[plan] }] and imports the same cusProduct.
+ * dfu.flash — Stripe recurring plan applies mid-cycle usage (contract 1,2,3):
+ * route resolves, a full v1 payload validates, and an active cusProduct is
+ * inserted with balance = allowance - usage.
  */
 
 import { expect, test } from "bun:test";
@@ -9,7 +10,7 @@ import {
 	type FlashClient,
 	callFlash,
 	createRealStripeSub,
-} from "@tests/integration/billing/dfu/dfu-flash/utils/flashTestUtils.js";
+} from "@tests/integration/billing/dfu/billing-import/utils/flashTestUtils.js";
 import { expectCustomerProducts } from "@tests/integration/billing/utils/expectCustomerProductCorrect.js";
 import { expectBalanceCorrect } from "@tests/integration/utils/expectBalanceCorrect.js";
 import { TestFeature } from "@tests/setup/v2Features.js";
@@ -19,11 +20,11 @@ import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
 import chalk from "chalk";
 
 test.concurrent(
-	`${chalk.yellowBright("dfu.flash: `plan` shorthand imports same cusProduct as phases form")}`,
+	`${chalk.yellowBright("dfu.flash: Stripe recurring plan applies mid-cycle usage")}`,
 	async () => {
-		const customerId = "dfu-flash-plan-shorthand";
+		const customerId = "dfu-flash-stripe-recurring";
 		const pro = products.pro({
-			id: "dfu-plan-shorthand-pro",
+			id: "dfu-pro",
 			items: [items.monthlyMessages({ includedUsage: 100 })],
 		});
 
@@ -43,20 +44,33 @@ test.concurrent(
 				{
 					processor: "stripe",
 					link: { subscription_id: subscriptionId },
-					// Singular `plan` — normalized to phases:[{ starts_at:"now", plans:[plan] }].
-					plan: {
-						plan_id: pro.id,
-						status: "active",
-						balances: [{ feature_id: TestFeature.Messages, usage: 40 }],
-					},
+					phases: [
+						{
+							starts_at: "now",
+							plans: [
+								{
+									plan_id: pro.id,
+									status: "active",
+									balances: [{ feature_id: TestFeature.Messages, usage: 40 }],
+								},
+							],
+						},
+					],
 				},
 			],
 		};
 
 		const flashRes = await callFlash(autumnV2_2 as FlashClient, payload);
-		expect(flashRes.errorCode).not.toBe("invalid_inputs");
-		expect(flashRes.errorCode).not.toBe("invalid_request");
 
+		// ── Contract 1: route resolves (not 404). Green now (stub 501) & post-impl (200). ──
+		expect(flashRes.errorCode).not.toBe("not_found");
+		expect(flashRes.errorMessage ?? "").not.toContain("status 404");
+
+		// ── Contract 2: full v1 payload validates (no 400 / zod reject). ──
+		expect(flashRes.errorCode).not.toBe("invalid_request");
+		expect(flashRes.errorCode).not.toBe("invalid_inputs");
+
+		// ── Contract 3: active cusProduct inserted; balance = allowance - usage. ──
 		const customer = await autumnV2_3.customers.get<ApiCustomerV5>(customerId);
 		await expectCustomerProducts({ customer, active: [pro.id] });
 		expectBalanceCorrect({
